Lagometer: Fix exception
This commit fixes an exception when using the Lagometer with scripts
that use the global variable Game in a way that the Lagometer isn't
expecting. I initially encountered this exception while using a self
contained module loaded through ScriptHawk alongside the Lagometer. It
might be wise to check for this type of exception in more places.