Beta: DK64 Temporary Flags
Don't know if the start of the block is correct. The start of the block was chosen as the first memory address after the last meaningful one with a known purpose that didn't give off any false readings with this script. There's a pointer pointing to 0x7FDDA0, hence 0x7FDD9F was chosen as the end.

Also not sure how to compress the script into the more optimised flag set/clear detection system that the main DK64 flag lua has