BT: Cleanup & Better floor vert detection
Cleanup of the player object stuff (it really annoyed me)

Addition of some better detection of floor verts (Previous system couldn't detect floor verts of large maps like JRL main or floor verts of objects. This system can). The downside is that the floor verts only update when you are within ~1000 units (vertically) of the floor